dev top4java 2.0

+++ src/com/taobao/top/request/ItemextrasSearchRequest.java
@@ -18,8 +18,8 @@ public class ItemextrasSearchRequest implements TopRequest {
 	private String numIids;
 	private Long options;
 	private String orderBy;
-	private Long pageNo;
-	private Long pageSize;
+	private Integer pageNo;
+	private Integer pageSize;
 	private String q;
 	private String sellerCids;
 	private Long shopId;
@@ -44,10 +44,10 @@ public class ItemextrasSearchRequest implements TopRequest {
 	public void setOrderBy(String orderBy) {
 		this.orderBy = orderBy;
 	}
-	public void setPageNo(Long pageNo) {
+	public void setPageNo(Integer pageNo) {
 		this.pageNo = pageNo;
 	}
-	public void setPageSize(Long pageSize) {
+	public void setPageSize(Integer pageSize) {
 		this.pageSize = pageSize;
 	}
 	public void setQ(String q) {
