+++ src/com/taobao/top/request/ItemsInventoryGetRequest.java
@@ -20,8 +20,8 @@ public class ItemsInventoryGetRequest implements TopRequest {
 	private Boolean isEx;
 	private Boolean isTaobao;
 	private String orderBy;
-	private Long pageNo;
-	private Long pageSize;
+	private Integer pageNo;
+	private Integer pageSize;
 	private String q;
 	private String sellerCids;
 
@@ -49,10 +49,10 @@ public class ItemsInventoryGetRequest implements TopRequest {
 	public void setOrderBy(String orderBy) {
 		this.orderBy = orderBy;
 	}
-	public void setPageNo(Long pageNo) {
+	public void setPageNo(Integer pageNo) {
 		this.pageNo = pageNo;
 	}
-	public void setPageSize(Long pageSize) {
+	public void setPageSize(Integer pageSize) {
 		this.pageSize = pageSize;
 	}
 	public void setQ(String q) {
