+++ src/com/taobao/top/request/TraderatesGetRequest.java
@@ -13,8 +13,8 @@ import com.taobao.top.util.TopHashMap;
 public class TraderatesGetRequest implements TopRequest {
 
 	private String fields;
-	private Long pageNo;
-	private Long pageSize;
+	private Integer pageNo;
+	private Integer pageSize;
 	private String rateType;
 	private String result;
 	private String role;
@@ -22,10 +22,10 @@ public class TraderatesGetRequest implements TopRequest {
 	public void setFields(String fields) {
 		this.fields = fields;
 	}
-	public void setPageNo(Long pageNo) {
+	public void setPageNo(Integer pageNo) {
 		this.pageNo = pageNo;
 	}
-	public void setPageSize(Long pageSize) {
+	public void setPageSize(Integer pageSize) {
 		this.pageSize = pageSize;
 	}
 	public void setRateType(String rateType) {
